requirements.txt
# AI Bridge Platform - Production Requirements

# Core Flask framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
bcrypt==4.1.2

# Redis for caching and sessions
redis==5.0.1
Flask-Session==0.5.0

# AWS S3 integration
boto3==1.34.0
botocore==1.34.0

# HTTP requests for Label Studio integration
requests==2.31.0

# Background tasks
celery==5.3.4

# Production server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0

# Data validation
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# Email support
Flask-Mail==0.9.1

# Monitoring and logging
sentry-sdk[flask]==1.38.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# File handling
Pillow==10.1.0

# Development tools (remove in production if needed)
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

